requirements.txt
pandas==1.5.3
requests==2.31.0
matplotlib==3.7.1
python-binance==1.0.17
pandas-ta==0.3.14b0